@@ -29,7 +29,7 @@ export const CommandExecutorLive = Layer.succeed(
   CommandExecutor.of({
     spawn: (command, ...args) =>
       Effect.gen(function*(_) {
-        const { status, stderr } = yield* _(Effect.try({
+        const { status, stderr, stdout } = yield* _(Effect.try({
           try: () => NodeChildProcess.spawnSync(command, args, { stdio: "pipe", encoding: "utf8" }),
           catch: (error) =>
             new Error(
@@ -38,10 +38,12 @@ export const CommandExecutorLive = Layer.succeed(
               }`
             )
         }))
+
         if (status !== 0) {
+          const output = [stdout, stderr].filter((x) => Boolean(x)).join("\n")
           return yield* _(Effect.fail(
             new Error(
-              `[CommandExecutor] During execution of '${command}', the following error(s) occurred:\n${stderr}`
+              `[CommandExecutor] During execution of '${command}', the following error(s) occurred:\n${output}`
             )
           ))
         }
